+++ MyInvestAPI/Controllers/UserController.cs
@@ -5,6 +5,7 @@ using MyInvestAPI.Domain;
 using MyInvestAPI.Repositories.Interfaces;
 using MyInvestAPI.Services.Interfaces;
 using MyInvestAPI.ViewModels;
+using System.Security.Claims;
 
 namespace MyInvestAPI.Controllers
 {
@@ -44,6 +45,12 @@ namespace MyInvestAPI.Controllers
         [HttpGet("{userId}", Name ="GetUser")]
         public async Task<ActionResult<User>> GetById(string userId)
         {
+            var userIdFromToken = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userIdFromToken != userId)
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, new { message = "Você não pode buscar outro usuário por id."});
+            }
+
             return Ok(await _repository.GetByIdAsync(userId));
         }
 
@@ -51,6 +58,12 @@ namespace MyInvestAPI.Controllers
         [HttpGet("{userId}/purses")]
         public async Task<ActionResult<User>> GetUserWithAllPursesById(string userId)
         {
+            var userIdFromToken = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userIdFromToken != userId)
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, new { message = "Você não pode buscar outro usuário por id." });
+            }
+
             return Ok(await _repository.GetUserWithAllPursesByIdAsync(userId));
         }
 
@@ -58,6 +71,12 @@ namespace MyInvestAPI.Controllers
         [HttpGet("{userId}/purses/actives")]
         public async Task<ActionResult<User>> GetUserWithAllPursesAndActivesById(string userId)
         {
+            var userIdFromToken = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userIdFromToken != userId)
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, new { message = "Você não pode buscar outro usuário por id." });
+            }
+
             return Ok(await _repository.GetUserWithAllPursesAndActivesByIdAsync(userId));
         }
 
@@ -65,6 +84,12 @@ namespace MyInvestAPI.Controllers
         [HttpPut("{userId}")]
         public async Task<IActionResult> UpdateAsync(string userId, CreateUserViewModel userViewModel)
         {
+            var userIdFromToken = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userIdFromToken != userId)
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, new { message = "Você não pode atualizar outro usuário que não seja você." });
+            }
+
             if (userViewModel is null)
                 return BadRequest("The data for update must not be null.");
 
@@ -77,6 +102,12 @@ namespace MyInvestAPI.Controllers
         [HttpDelete("{userId}")]
         public async Task<IActionResult> Delete(string userId)
         {
+            var userIdFromToken = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (userIdFromToken != userId)
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, new { message = "Você não pode deletar outro usuário que não seja você." });
+            }
+
             await _repository.Delete(userId);
             return NoContent();
         }
